chore: make root aliases a global barrel file
