--- lower_upper.go
@@ -1,1 +1,0 @@
-package flect
